Move code of most portal buttons to the portal module.
